+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-08-31 03:00:16
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 18f50e6b31e038d8919fb012132aa03478d9e4c2
+
+## [ForwardWidgets-2kuai] - 2025-08-31 03:00:16
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: c067734ff745951e7c97bfb645f1fdbf87e75847
+
+## [ForwardWidgets-pack1r] - 2025-08-31 03:00:16
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 185e88999e8e4b7cdebb3885c4ca8d730ac8be9b
+
+
 ## [ForwardWidgets-huangxd] - 2025-08-30 02:48:39
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main